backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
yfinance>=0.2.48
pandas
numpy
fredapi>=0.5.1
feedparser>=6.0.11
groq>=0.11.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
